Refactor: Isolate CGEnum into a separate module for build infrastructure optimization
- Extract CGEnum class from monolithic codegen.py to enum_generator.py
- Ensure strict module isolation and reduce coupling in WebIDL binding generation
- Improve readability by converting multi-line string code generation into structured layouts
